requirements.txt
openai
python-dotenv
pillow
easyocr
flask             # if app.py is a web server
